+++ src/utils/withdrawals.ts
@@ -6,10 +6,6 @@
 export const getWithdrawalVaultUUID = (uuid: bigint) =>
   uuid & BigInt(0x00000000ffffffff);
 
-/**
- * Gets the Vault index associated to the Withdrawal.
- * @param uuid The withdrawal UUID
- * @returns The vault UUID.
- */
+/** @knipignore */
 export const getVaultUUID = (uuid: bigint) =>
   (uuid & BigInt(0xffffffff00000000)) >> BigInt(32);
